+++ consts.py
@@ -1,0 +1,5 @@
+
+data_constant = {"screen_width": 1000, "screen_height": 800, "radius": 300, "circle_color": (255, 255, 255),
+                 "circle_width": 1, "player_speed": 5, "arrow_speed": 1}
+
+data_constant["center"] = (data_constant["screen_width"] // 2, data_constant["screen_height"] // 2)
